LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/084-expected.txt

CONSOLE MESSAGE: line 1: TypeError: null is not an object (evaluating 'top.log')
FAILED (This TC requires JavaScript enabled)

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'top.log')
